Module aws_sdk_translate::error

Modules

concurrent_modification_exception

See ConcurrentModificationException

conflict_exception

See ConflictException

detected_language_low_confidence_exception

See DetectedLanguageLowConfidenceException

internal_server_exception

See InternalServerException

invalid_filter_exception

See InvalidFilterException

invalid_parameter_value_exception

See InvalidParameterValueException

invalid_request_exception

See InvalidRequestException

limit_exceeded_exception

See LimitExceededException

resource_not_found_exception

See ResourceNotFoundException

service_unavailable_exception

See ServiceUnavailableException

text_size_limit_exceeded_exception

See TextSizeLimitExceededException

too_many_requests_exception

See TooManyRequestsException

unsupported_language_pair_exception

See UnsupportedLanguagePairException

Structs

ConcurrentModificationException

Another modification is being made. That modification must complete before you can make

ConflictException

There was a conflict processing the request. Try your request again.

CreateParallelDataError

DeleteParallelDataError

DeleteTerminologyError

DescribeTextTranslationJobError

DetectedLanguageLowConfidenceException

The confidence that Amazon Comprehend accurately detected the source language is low. If a

GetParallelDataError

GetTerminologyError

ImportTerminologyError

InternalServerException

An internal server error occurred. Retry your request.

InvalidFilterException

The filter specified for the operation is invalid. Specify a different filter.

InvalidParameterValueException

The value of the parameter is invalid. Review the value of the parameter you are using to

InvalidRequestException

The request that you made is invalid. Check your request to determine why it's invalid

LimitExceededException

The specified limit has been exceeded. Review your request and retry it with a quantity

ListParallelDataError

ListTerminologiesError

ListTextTranslationJobsError

ResourceNotFoundException

The resource you are looking for has not been found. Review the resource you're looking

ServiceUnavailableException

The Amazon Translate service is temporarily unavailable. Please wait a bit and then retry your

StartTextTranslationJobError

StopTextTranslationJobError

TextSizeLimitExceededException

The size of the text you submitted exceeds the size limit. Reduce the size of the text or

TooManyRequestsException

You have made too many requests within a short period of time. Wait for a short time and

TranslateTextError

UnsupportedLanguagePairException

Amazon Translate does not support translation from the language of the source text into the requested

UpdateParallelDataError

Enums

CreateParallelDataErrorKind

DeleteParallelDataErrorKind

DeleteTerminologyErrorKind

DescribeTextTranslationJobErrorKind

GetParallelDataErrorKind

GetTerminologyErrorKind

ImportTerminologyErrorKind

ListParallelDataErrorKind

ListTerminologiesErrorKind

ListTextTranslationJobsErrorKind

StartTextTranslationJobErrorKind

StopTextTranslationJobErrorKind

TranslateTextErrorKind

UpdateParallelDataErrorKind