Module aws_sdk_translate::model

Modules

applied_terminology

See AppliedTerminology

encryption_key

See EncryptionKey

input_data_config

See InputDataConfig

job_details

See JobDetails

output_data_config

See OutputDataConfig

parallel_data_config

See ParallelDataConfig

parallel_data_data_location

See ParallelDataDataLocation

parallel_data_properties

See ParallelDataProperties

term

See Term

terminology_data

See TerminologyData

terminology_data_location

See TerminologyDataLocation

terminology_properties

See TerminologyProperties

text_translation_job_filter

See TextTranslationJobFilter

text_translation_job_properties

See TextTranslationJobProperties

Structs

AppliedTerminology

The custom terminology applied to the input text by Amazon Translate for the translated text

EncryptionKey

The encryption key used to encrypt this object.

InputDataConfig

The input configuration properties for requesting a batch translation job.

JobDetails

The number of documents successfully and unsuccessfully processed during a translation

OutputDataConfig

The output configuration properties for a batch translation job.

ParallelDataConfig

Specifies the format and S3 location of the parallel data input file.

ParallelDataDataLocation

The location of the most recent parallel data input file that was successfully imported

ParallelDataProperties

The properties of a parallel data resource.

Term

The term being translated by the custom terminology.

TerminologyData

The data associated with the custom terminology.

TerminologyDataLocation

The location of the custom terminology data.

TerminologyProperties

The properties of the custom terminology.

TextTranslationJobFilter

Provides information for filtering a list of translation jobs. For more information, see

TextTranslationJobProperties

Provides information about a translation job.

Enums

EncryptionKeyType

JobStatus

MergeStrategy

ParallelDataFormat

ParallelDataStatus

TerminologyDataFormat