Struct aws_sdk_translate::model::text_translation_job_properties::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for TextTranslationJobProperties

Implementations

impl Builder

pub fn job_id(self, input: impl Into<String>) -> Self

The ID of the translation job.

pub fn set_job_id(self, input: Option<String>) -> Self

pub fn job_name(self, input: impl Into<String>) -> Self

The user-defined name of the translation job.

pub fn set_job_name(self, input: Option<String>) -> Self

pub fn job_status(self, input: JobStatus) -> Self

The status of the translation job.

pub fn set_job_status(self, input: Option<JobStatus>) -> Self

pub fn job_details(self, input: JobDetails) -> Self

The number of documents successfully and unsuccessfully processed during the translation job.

pub fn set_job_details(self, input: Option<JobDetails>) -> Self

pub fn source_language_code(self, input: impl Into<String>) -> Self

The language code of the language of the source text. The language must be a language supported by Amazon Translate.

pub fn set_source_language_code(self, input: Option<String>) -> Self

pub fn target_language_codes(self, input: impl Into<String>) -> Self

pub fn set_target_language_codes(self, input: Option<Vec<String>>) -> Self

pub fn terminology_names(self, input: impl Into<String>) -> Self

pub fn set_terminology_names(self, input: Option<Vec<String>>) -> Self

pub fn parallel_data_names(self, input: impl Into<String>) -> Self

pub fn set_parallel_data_names(self, input: Option<Vec<String>>) -> Self

pub fn message(self, input: impl Into<String>) -> Self

An explanation of any errors that may have occured during the translation job.

pub fn set_message(self, input: Option<String>) -> Self

pub fn submitted_time(self, input: Instant) -> Self

The time at which the translation job was submitted.

pub fn set_submitted_time(self, input: Option<Instant>) -> Self

pub fn end_time(self, input: Instant) -> Self

The time at which the translation job ended.

pub fn set_end_time(self, input: Option<Instant>) -> Self

pub fn input_data_config(self, input: InputDataConfig) -> Self

The input configuration properties that were specified when the job was requested.

pub fn set_input_data_config(self, input: Option<InputDataConfig>) -> Self

pub fn output_data_config(self, input: OutputDataConfig) -> Self

The output configuration properties that were specified when the job was requested.

pub fn set_output_data_config(self, input: Option<OutputDataConfig>) -> Self

pub fn data_access_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that granted Amazon Translate read access to the job's input data.

pub fn set_data_access_role_arn(self, input: Option<String>) -> Self

pub fn build(self) -> TextTranslationJobProperties

Consumes the builder and constructs a TextTranslationJobProperties