Module aws_sdk_translate::operation

Structs

CreateParallelData

Creates a parallel data resource in Amazon Translate by importing an input file from

DeleteParallelData

Deletes a parallel data resource in Amazon Translate.

DeleteTerminology

A synchronous action that deletes a custom terminology.

DescribeTextTranslationJob

Gets the properties associated with an asycnhronous batch translation job including name,

GetParallelData

Provides information about a parallel data resource.

GetTerminology

Retrieves a custom terminology.

ImportTerminology

Creates or updates a custom terminology, depending on whether or not one already exists

ListParallelData

Provides a list of your parallel data resources in Amazon Translate.

ListTerminologies

Provides a list of custom terminologies associated with your account.

ListTextTranslationJobs

Gets a list of the batch translation jobs that you have submitted.

StartTextTranslationJob

Starts an asynchronous batch translation job. Batch translation jobs can be used to

StopTextTranslationJob

Stops an asynchronous batch translation job that is in progress.

TranslateText

Translates input text from the source language to the target language. For a list of

UpdateParallelData

Updates a previously created parallel data resource by importing a new input file from